+++ interface/libs/commands/include/cmd/nrf24l01_init/command.hxx
@@ -12,7 +12,7 @@ public:
     {
         COMMANDS::NRF24L01_INIT::command_t command;
 
-        for (int i = 0; i < sizeof(command.txAddr); i++) {
+        for (uint8_t i = 0; i < sizeof(command.txAddr); i++) {
             if (i >= txAddr.size()) {
                 break;
             }
@@ -21,7 +21,7 @@ public:
                 = txAddr.at(i);
         }
 
-        for (int i = 0; i < sizeof(command.rxAddr); i++) {
+        for (uint8_t i = 0; i < sizeof(command.rxAddr); i++) {
             if (i >= rxAddr.size()) {
                 break;
             }
@@ -43,13 +43,13 @@ public:
     {
         COMMANDS::NRF24L01_INIT::command_t command;
 
-        for (int i = 0; i < txAddr.size() && i < 5; i++) {
+        for (uint8_t i = 0; i < txAddr.size() && i < 5; i++) {
             m_payload.at(
                 offsetof(COMMANDS::NRF24L01_INIT::command_t, txAddr[0]) + i)
                 = txAddr.at(i);
         }
 
-        for (int i = 0; i < rxAddr.size() && i < 5; i++) {
+        for (uint8_t i = 0; i < rxAddr.size() && i < 5; i++) {
             m_payload.at(
                 offsetof(COMMANDS::NRF24L01_INIT::command_t, rxAddr[0]) + i)
                 = rxAddr.at(i);
